requirements.txt
azure-identity==1.20.0
azure-ai-projects==1.0.0b11
python-dotenv
termcolor==2.4.0
mcp==1.9.2